Implement command execution
next_special() - finds the index of the next special char
------------------------------------------------------------------
This function is implemented to help identify special chars in the
shell command and to assist in identifying a split point for those
commands.

executeCmd() - base function for recieving the shell arguements
------------------------------------------------------------------
Function determines whether to execute in background or foreground
then calls execute().

execute() - Parses the exectution string and executes commands
------------------------------------------------------------------
Uses the proper helper functions to assist in executing commands.
This function is responsible for setting up the proper file
pointers to handle redirects and piping.