Add multithread functionality to test program in main
Program splits the range of primes to search for based on the number
of threads allocated. The program then creates threads for each subset
of numbers and outputs these to a file called "prime*.txt" where * is
the thread number. The main thread then waits for all the worker
threads to complete and then concatenates all of the thread output
files together to create a file called "primes.txt".